+++ requirements.txt
@@ -1,5 +1,16 @@
+alabaster==0.7.6
+Babel==2.1.1
 Django==1.9
 django-debug-toolbar==1.4
+docutils==0.12
+Jinja2==2.8
+MarkupSafe==0.23
+Pygments==2.0.2
+pytz==2015.7
+requests==2.8.1
 six==1.10.0
+snowballstemmer==1.2.0
+Sphinx==1.3.3
+sphinx-rtd-theme==0.1.9
 sqlparse==0.1.18
 wheel==0.24.0
